pyproject.toml
[project]
name = "real-estate"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "asyncio>=3.4.3",
    "bs4>=0.0.2",
    "cassandra-driver>=3.29.2",
    "docker>=7.1.0",
    "ipykernel>=6.29.5",
    "kafka-python>=2.0.2",
    "numpy>=2.0.2",
    "openai>=1.51.2",
    "pandas>=2.2.3",
    "playwright>=1.47.0",
    "pyspark>=3.5.3",
    "python-dotenv>=1.0.1",
    "selenium-async>=0.1.0",
    "selenium>=4.25.0",
    "undetected-chromedriver>=3.5.5",
]
